# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projetos:
    get:
      operationId: listProjetos
      tags: [projetos]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Projeto"
    post:
      operationId: createProjeto
      tags: [projetos]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjetoBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projeto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projetos/{id}:
    get:
      operationId: getProjeto
      tags: [projetos]
      summary: Get a project by ID (with members and tasks)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjetoDetalhe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProjeto
      tags: [projetos]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjetoBody"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projeto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProjeto
      tags: [projetos]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Project deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Membros de Projeto ─────────────────────────────────────────
  /projetos/{projetoId}/membros:
    get:
      operationId: listProjetoMembros
      tags: [membros]
      summary: List members of a project
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjetoMembro"
    post:
      operationId: addProjetoMembro
      tags: [membros]
      summary: Add a member to a project
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProjetoMembroBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjetoMembro"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projetos/{projetoId}/membros/{usuarioId}:
    delete:
      operationId: removeProjetoMembro
      tags: [membros]
      summary: Remove a member from a project
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Tarefas ────────────────────────────────────────────────────
  /projetos/{projetoId}/tarefas:
    get:
      operationId: listTarefas
      tags: [tarefas]
      summary: List tasks for a project
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tarefa"
    post:
      operationId: createTarefa
      tags: [tarefas]
      summary: Create a task in a project
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTarefaBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarefa"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projetos/{projetoId}/tarefas/{id}:
    patch:
      operationId: updateTarefa
      tags: [tarefas]
      summary: Update a task
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTarefaBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarefa"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTarefa
      tags: [tarefas]
      summary: Delete a task
      parameters:
        - name: projetoId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ─────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Enums ──────────────────────────────────────────────────
    Perfil:
      type: string
      enum: [ADMIN, GERENTE, COLABORADOR]
    StatusProjeto:
      type: string
      enum: [PLANEJAMENTO, EM_ANDAMENTO, CONCLUIDO, CANCELADO]
    StatusTarefa:
      type: string
      enum: [PENDENTE, EM_ANDAMENTO, CONCLUIDA, CANCELADA]
    PrioridadeTarefa:
      type: string
      enum: [BAIXA, MEDIA, ALTA, CRITICA]

    # ── Usuario ────────────────────────────────────────────────
...
    Projeto:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/StatusProjeto"
        dataInicio:
          type: string
          format: date-time
        dataPrazo:
          type: string
          format: date-time
        gerenteId:
          type: integer
        gerenteNome:
          type: string
        totalMembros:
          type: integer
        totalTarefas:
          type: integer
        tarefasConcluidas:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nome
        - status
        - dataInicio
        - dataPrazo
        - gerenteId
        - gerenteNome
        - totalMembros
        - totalTarefas
        - tarefasConcluidas
        - createdAt
        - updatedAt
    ProjetoDetalhe:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/StatusProjeto"
        dataInicio:
          type: string
          format: date-time
        dataPrazo:
          type: string
          format: date-time
        gerenteId:
          type: integer
        gerenteNome:
          type: string
        membros:
          type: array
          items:
            $ref: "#/components/schemas/ProjetoMembro"
        tarefas:
          type: array
          items:
            $ref: "#/components/schemas/Tarefa"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nome
        - status
        - dataInicio
        - dataPrazo
        - gerenteId
        - gerenteNome
        - membros
        - tarefas
        - createdAt
        - updatedAt
    CreateProjetoBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        status:
          $ref: "#/components/schemas/StatusProjeto"
        dataInicio:
          type: string
          format: date-time
        dataPrazo:
          type: string
          format: date-time
        gerenteId:
          type: integer
      required: [nome, status, dataInicio, dataPrazo, gerenteId]
    UpdateProjetoBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        status:
          $ref: "#/components/schemas/StatusProjeto"
        dataInicio:
          type: string
          format: date-time
        dataPrazo:
          type: string
          format: date-time
        gerenteId:
          type: integer
...
    ProjetoMembro:
      type: object
      properties:
        id:
          type: integer
        projetoId:
          type: integer
        usuarioId:
          type: integer
        usuarioNome:
          type: string
        usuarioCargo:
          type: string
        usuarioPerfil:
          $ref: "#/components/schemas/Perfil"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projetoId
        - usuarioId
        - usuarioNome
        - usuarioCargo
        - usuarioPerfil
        - createdAt
    AddProjetoMembroBody:
      type: object
      properties:
        usuarioId:
          type: integer
      required: [usuarioId]

    # ── Tarefa ─────────────────────────────────────────────────
    Tarefa:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descricao:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/StatusTarefa"
        prioridade:
          $ref: "#/components/schemas/PrioridadeTarefa"
        projetoId:
          type: integer
        responsavelId:
          type: ["integer", "null"]
        responsavelNome:
          type: ["string", "null"]
        dataVencimento:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - titulo
        - status
        - prioridade
        - projetoId
        - createdAt
        - updatedAt
    CreateTarefaBody:
      type: object
      properties:
        titulo:
          type: string
        descricao:
          type: string
        status:
          $ref: "#/components/schemas/StatusTarefa"
        prioridade:
          $ref: "#/components/schemas/PrioridadeTarefa"
        responsavelId:
          type: integer
        dataVencimento:
          type: string
          format: date-time
      required: [titulo, status, prioridade]
    UpdateTarefaBody:
      type: object
      properties:
        titulo:
          type: string
        descricao:
          type: string
        status:
          $ref: "#/components/schemas/StatusTarefa"
        prioridade:
          $ref: "#/components/schemas/PrioridadeTarefa"
        responsavelId:
          type: integer
        dataVencimento:
          type: string
          format: date-time

    # ── Dashboard ──────────────────────────────────────────────
...